+++ changelog.md
@@ -5,6 +5,7 @@ v1.14.18:
 - Fixed Runic Power showing on a scale of 0-1000+ instead of 0-100+ with DogTags off
 - Added option (enabled by default) to hide mod during cataloging.
 - Fixed "Hide Blizzard Buffs" option on PlayerInfo causing errors and "?" icons when toggling off.
+- Also hide Debuff frame if it exists (Dragonflight+) when enabling "Hide Blizzard Buffs" in the PlayerInfo module.
 
 v1.14.17: